PostJob: do not do fake get_transfer call
The get_transfers_statuses was done outside the while loo to give
ASO a chance to update the documents. Now that's not needed anymore
since we check staleness of documents usint the start transfer time.

Note from @atanasi: check that if start_time patch is working in
PostJob, then we can remove the call to the view before starting
the loop, and move the sleep back to the end of the loop.