Remove the userFiles parameter in the job wrapper
These are the additional user files a user can set up in the client so
that they are added to the sandbox. Since those files are already in the
sandbox we do not need to move them around, and we do not need to track
them. FYI, this parameter was not even passed to the jobwrapper.