Align OS requiements with WMAgent. (#5460)
We have been moving the OS requirement lines out of the job's
REQUIREMENTS expression and into a separate attribute in the job.
(similar to RequiredCPUs).  This allows the submit infrastructure
layer to enforce the requirements on the pilot-side, meaning less
code and deployment coordination is necessary across all three teams
for major changes.

As an example, the existing GLIDEIN_REQUIRED_OS expression does not
work with Singularity-inside-RHEL7 (luckily, it works with Singularity-
inside-RHEL6).  We should not specify it at the CRAB level and let
SI handle this.