Fix resubmission with automatic splitting
When using the external SubDAG line of DAGMan, HTCondor actively removes
remaining jobs in the queue when aborting a DAG due to failed nodes. This
includes tail SubDAGs and makes resubmission impossible. Make the DAGs
independent by submitting them when we previously created a submission file
only, and switch the DAG nodes to NOOP.